% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_plot_AR.R
\name{CAP_plot_AR}
\alias{CAP_plot_AR}
\title{Function plots CAP curve, and calculate Accuracy Rate of the given logistic regression model.}
\usage{
CAP_plot_AR(logistic_probability_vector, given_outcome_vector_from_data)
}
\arguments{
\item{logistic_probability_vector}{Vector of probabilities predicted by logistic regression model.}

\item{given_outcome_vector_from_data}{Vector of outcomes from train or test data.}
}
\value{
CAP curve and Accuracy rate of the model.
}
\description{
Function plots CAP curve, and calculate Accuracy Rate of the given logistic regression model.
}
\examples{
CAP_plot_AR(predicted,outcome_from_data)
}
